Move quiz_worker to root dir
