+++ lib/slaq.rb
@@ -1,6 +1,7 @@
 require_relative 'slaq/answer'
 require_relative 'slaq/version'
 require_relative 'slaq/quiz'
+require_relative 'slaq/slack'
 require_relative 'slaq/redis'
 require_relative 'slaq/wikipedia'
 require_relative 'slaq/command'
